Update image and move landing page images into one folder
