+++ packages/website/src/assets/img/landing-page/buoy.jpg
