+++ packages/website/src/assets/img/landing-page/metrics.png
