+++ packages/website/src/assets/img/landing-page/survey-images.png
